asoc: lahaina: Add ext clk source support
For audio interfaces to be sourced from an external clock, one needs to
configure clock attributes such as the clock root and RCG registers for
operation at an appropriate clock frequency. Add support for this by
registering for a callback with Q6AFE driver which during clock enable,
provides clock attribute values and toggles the GPIO pin sourcing the
external clock.

Change-Id: I64a849e045a13dedd876b6a82c727a0e0610cbb7
Signed-off-by: Abhilash Sathyanarayana <quic_abhisath@quicinc.com>