qcacld-3.0: Update key management in original auth mode for WAPI
Currently for WAPI connection auth mode is updated only
in key management, but on connect request for validating
auth mode, it compared with original key management and
failing to connect because of mismatch in auth and cipher suits.

To fix this, update auth mode in original key management as well
in crypto for WAPI connection.

CRs-Fixed: 3901275
Change-Id: I942fbbc68c18d7c8137053b5fc9cb0260109fdcd